feat: 0.5학점 입력 + 대시보드 오표시 제거 + 전공 평가특성 30과목 보강
[0.5학점] Course.credit int → double. 인성과미래설계Ⅰ(0.5학점)을 credit=0으로
우회하던 합의를 해제하고 실제 값으로 입력 가능하게 함.
- 요건 기준값(130·21·39)은 정수라 required는 int 유지, 과목 합계인
  earned·shortfall만 double.
- GpaCalculator: int gpaCredit += double 복합대입이 암묵적 축소 캐스팅으로
  0.5를 조용히 잘라 GPA를 틀리게 하던 버그 수정(컴파일 에러 없이 통과했음).
- CourseService.validateCredit: int 타입이 하던 가드를 대체 — 0~30, 0.5 단위만.
- H2는 ddl-auto=update가 alter column credit set data type float(53)을 실행해
  기존 데이터를 무손실 마이그레이션(재기동만으로 적용).

[오표시] 졸업시험/논문이 판정 데이터가 없는데 status="완료"로 나가던 문제.
BIGDATA_AI는 thesisCertRequired=false → graduationExam=NONE → "완료" 판정이었음.
THESIS·EITHER만 판정하고 나머지는 "미입력". 핵심교양도 "대상 영역 수"를
진척도로 내보내 5/5 "완료"가 되던 것을 current/required=null + "미입력"으로.

[평가특성] 전공 평가특성 11 → 30과목. 원석 제공 CSV·강의평가표 반영,
⑧성적후함정도 빈칸은 보통으로, 데이터 부재 5건은 과목 성격 기반 임의 추정치
(비고에 근거·추정 여부 명시). 졸업논문은 P/NP라 8개 분류원칙 대상이 아니므로
빈칸 유지(persona·방향분류에서 자동 제외).

[build_data] 평가특성 키를 교과목코드 → (교과목명, 교과목코드) 쌍으로 변경.
코드만으로는 정밀의료/BM프로젝트(400990), K-콘텐츠새로읽기/대인관계론(401441),
웰니스문화의이해/광고와문화(400778)가 서로를 덮어써 엉뚱한 특성이 붙었다.
이름만으로는 실용영어청취및말하기(400301/114511)가 겹친다. 쌍이면 182/182 1:1.

⚠️ 지문에 credit이 포함되어(3 → 3.0) 기존 추천 캐시가 전부 무효화됨.
   다음 POST /recommendations에서 재생성됨.

+++ src/main/java/com/passport/course/domain/Course.java
@@ -27,7 +27,7 @@ public class Course {
     private String name;
 
     @Column(nullable = false)
-    private int credit;
+    private double credit;
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
@@ -48,7 +48,7 @@ public class Course {
     private boolean retake;
 
     @Builder
-    public Course(Profile profile, String name, int credit, CourseCategory category, Grade grade, int year, int semester, boolean retake) {
+    public Course(Profile profile, String name, double credit, CourseCategory category, Grade grade, int year, int semester, boolean retake) {
         this.profile = profile;
         this.name = name;
         this.credit = credit;
@@ -59,7 +59,7 @@ public class Course {
         this.retake = retake;
     }
 
-    public void update(String name, int credit, CourseCategory category, Grade grade, int year, int semester, boolean retake) {
+    public void update(String name, double credit, CourseCategory category, Grade grade, int year, int semester, boolean retake) {
         this.name = name;
         this.credit = credit;
         this.category = category;
