+++ src/main/java/com/passport/course/dto/CourseCreateRequest.java
@@ -4,7 +4,7 @@ import com.passport.course.domain.Course;
 
 public record CourseCreateRequest(
         String name,
-        int credit,
+        double credit,
         Course.CourseCategory category,
         Course.Grade grade,
         int year,
