+++ src/main/java/com/passport/dashboard/dto/DashboardResponse.java
@@ -25,8 +25,8 @@ public record DashboardResponse(
         List<CategoryView> categories,
         /** 최신 학기("YYYY-N학기"). 수강 이력이 전혀 없으면 null */
         String semester,
-        int requestedCredits,
-        int earnedCredits,
+        double requestedCredits,
+        double earnedCredits,
         /** 최신 학기 GPA. 기존 최상위 gpa(GpaProgress 객체)와 이름이 겹쳐 semesterGpa로 명명 — ⚠️ FE 확인 필요 */
         Double semesterGpa,
         /** 최신 학기 수강목록 */
@@ -42,13 +42,15 @@ public record DashboardResponse(
     public record CategoryView(
             String key,
             String name,
-            Integer current,
+            /** 학점 항목은 0.5 단위가 있어 소수(예: 39.5). 건수·영역수 항목도 같은 타입으로 내려간다(3.0 → JS에선 3). */
+            Double current,
+            /** 요건 기준값·건수는 항상 정수. 판정 근거가 없으면 null. */
             Integer required,
             @JsonInclude(JsonInclude.Include.NON_NULL) String unit,
             String status
     ) {
     }
 
-    public record CourseLine(String category, String courseName, int credit, String grade) {
+    public record CourseLine(String category, String courseName, double credit, String grade) {
     }
 }
